Check the duplicated positional dispatch for drift, and split the page-composition diff
Two things I left as known gaps last time.

The positional dispatch is written twice — markup-core-api/main.mbt and
markup-core-cli/main.mbt, 61 commands each, 1,487 lines that must agree
with nothing checking they did. The failure is quiet by construction: both
backends are reachable at runtime, so a command implemented in only one of
them is correct until something falls back, then wrong with no error.

The test works at two levels and says so rather than implying more:
command names in full, which is the whole of the "added to one file,
forgotten in the other" class; and behaviour on five invocations chosen
for shape, including the 36-argument one where a swapped pair is least
likely to be caught by reading. Confirmed it fails on a simulated rename
rather than passing vacuously, and its extractor has the same
"did I match anything" guard as the workflow-command test.

Writing it turned up a cost the design doc had missed. The direct entry
point takes one tab-joined string, so an empty argument is
unrepresentable — ["cmd", ""] and ["cmd"] encode identically — and
production carries a __VLMKIT_EMPTY_ARG__ sentinel to work around it. My
first version of the helper skipped the sentinel, which made one command
return a usage message from one backend and an empty string from the
other; that looked like a backend asymmetry and was the helper not
speaking the protocol.

page-compose.ts also split: 473 lines of bounding-box arithmetic and
Markdown out of a file that also loads PNGs, drives Playwright and writes
crops. Unlike component-from-image.ts there is no import-cost win to claim
here — this file reaches Playwright through a dynamic import, so it never
paid one. The justification is narrower and worth stating plainly: a
spatial-pairing bug no longer needs a rendered page to reproduce, and a
reader looking for the algorithm finds a file about pairing rather than one
that also shells out to a browser. The CLI entry point keeps its path and
re-exports, so `vlmkit build page`, the MCP tool and existing imports are
untouched.
